[open][neva_upstreamable][app_runtime] Fix use-after-free because of typo in app_runtime WebAppWindow
:Release Notes:
Fix use-after-free because of typo assigning native widget pointer in app_runtime::WebAppWindow::InitWindow

:Detailed Notes:
For some reason we were using a local variable desktop_native_widget_aura_
in app_runtime::WebAppWindow::InitWindow instead of assigning to the
property of the same name. So when WebAppWindow was destroyed, the
native widget was not notified accordingly.

That causes a use-after-free on closing web apps, as the native widget
would try to notify the window it was going to be destroyed, after the
window was already destroyed.

:Testing Performed:
Builds, runs.

:QA Notes:
Test-cases: N/A

:Issues Addressed:

Change-Id: I76df8d036e97719c2bf798e71ee6297c235b1622
Originally-Reviewed-on: http://gpro.lgsvl.com/233420
Reviewed-on: http://gpro.lgsvl.com/233428
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>