[open][neva_upstreamable][ime] Unregister window from text input on destroy.
:Release Notes:
Unregister window from text input when it is destroyed.

:Detailed Notes:
WebOS text input keeps track of the active window and last active window.
So when a window is destroyed, we need to make sure the window is not
registered in text input. Otherwise, in certain text input wayland calls
we could be trying to access it.

:Testing Performed:
Valgrind was crashing on closign settings app. With
the fix it does not crash anymore.

:QA Notes:
Test-cases: N/A

:Issues Addressed:

Change-Id: I79005ff605ca50a22337faae07430fdd721d2257
Originally-Reviewed-on: http://gpro.lgsvl.com/233422
Reviewed-on: http://gpro.lgsvl.com/233430
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>