fixup! [ose][downstream][acg] Migrate chromium68 to ACG
:Release Notes:
Granted access to the 'Settings' service for the 'app-shell' binary

:Detailed Notes:
To provide sync-up of the 'navigator.language' property with the current system
locale, corresponding Luna Bus role as well as client permissions entities for
the 'app-shell' executable need to be added.

What's done:

- added outbound permission entity into the 'app-shell.role.json' file to
  access 'com.webos.settingsservice'
- added the 'Settings' service read-only permission entity into the
  'app-shell.perm.json' file

:Testing Performed:
Start building the webOS OSE build and make sure it's successfully created.

:QA Notes:
N/A

:Issues Addressed:
[PLAT-68386] [webOS OSE] navigator.language is not the same value with system's
             language

Change-Id: Ib4fd107e51c096d4b1b2b5064eb8dc4fbf75a012
Reviewed-on: http://gpro.lgsvl.com/233540
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Reviewed-by: Jose Dapena <jose.dapena@lge.com>