[open][neva_upstreamable][webos] Move WebOSLunaService to Neva.
:Release Notes:
Refactor WebOSLunaService moving it to neva.

:Detailed Notes:
This is a refactor of WebOSLunaService, to avoid reverse dependencies,
and being able to initialize it in the right time (browser process,
before main message loop is started).

We found that zygote was launching WebOSLunaservice, that is expected
to be available only in browser process. The initial luna connection
was launched in zygote process, and it could be behind the random
crashes we get in that same connection because it is available
in all the renderer processes.

:Testing Performed:
Stress tests: we should avoid crashes in renderer process in strlen
called from _LSTransportMessageGetPayload. We should also avoid the
more frequently seen PmLog error "Received message of size X bytes"
being X a huge number.

:QA Notes:
Test-cases: stress_test_15h

:Issues Addressed:
[PLAT-75337] Stress tests crash: strlen from _LSTransportMessageGetPayload

Change-Id: I44758efda524ac0479c2cb0f872dc87d12236f33
Originally-Reviewed-on: http://gpro.lgsvl.com/233761
Reviewed-on: http://gpro.lgsvl.com/233771
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>