[open][neva_upstreamable][media] Update display window when only source rect is changed in UMS playback.
:Release Notes:
Update display window when only source rect is changed in UMS playback.

:Detailed Notes:
Problem:
Previously, we changed our computation logic for display window to not
update display window if only source rect is changed. Because we assume
that platform media backend will realize such case and adjust by itself
automatically. But this assumption is valid only for MSE playback.
In UMS playback, we still need to update display window in such case.

Solution:
In UMS playback, update display window if source rect is changed.

Note: We may consider change UpdateVideoHoleBoundary() directly, but
this approach is avoided because the function is candidate for unifying
between WebMediaPlayerNeva and WebMediaPlayerMSE.

:Testing Performed:
1. Launch STARZ application.
2. Play any trailer video.
3. The video should be displayed right after playing started.
   (Before fix, the video was displayed after 10 seconds.)

:QA Notes:
Test-cases: GT-27189

:Issues Addressed:
[PLAT-75920] Video is not displayed during first 10 seconds.
             (audio is heard)

Change-Id: I98fd544931c472612af359aa0cb9148babf8cdc8
Originally-Reviewed-on: http://gpro.lgsvl.com/233826
Reviewed-on: http://gpro.lgsvl.com/233874
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>