[open][neva_upstreamable][fonts] Match CSS local font faces with style/stretch/weight
:Release Notes:
Match CSS local font faces with style/stretch/weight

:Detailed Notes:
Blink local font face normalizes the fonts requests to ignore the
font description attributes, like weight or style. It adds an
exception for Android to match properly the bold and thin flavours
of the Roboto font.

We have the same problem in webOS to match properly the variants of
Museo Sans. This patch avoids the normalization also in webOS.

:Testing Performed:
1. Set device in Korean
2. Go to Settings/General, and see with inspector the font of
  "(Language)" string.
Result: with fix it is "Museo Sans". Without the fix it was "Museo
Sans Medium".

1. Set device to English
2. Go to Settings
Result: left toolbar labels are italic now. Combo-like entries mix
bold and thin now.

:QA Notes:
Test-cases: GT-18893

:Issues Addressed:
[PLAT-76009] The font differs from the expected result.

Change-Id: I1397b1692a6841abf763de9cd2eb158df44ec409
Originally-Reviewed-on: http://gpro.lgsvl.com/234642
Reviewed-on: http://gpro.lgsvl.com/234828
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>