[open][neva_upstreamable][fonts] Fix typo passing font style queries to browser process
:Release Notes:
Fix typo passing bold/italic values in font queries to browser process.

:Detailed Notes:
Renderer process font queries are forwarded through sandbox IPC to
browser process. But the values for bold and italic were swapped when
populating the parameters to send to Sandbox IPC. This sets the proper
order of the arguments.

This patch is set as downstream, as this code has been replaced
in upstream with a different one that does not have the problem (it
passes explicitely booleans for is_bold and is_italic instead of
mixing them as bits in an int).

:Testing Performed:
With GDB, font with weight 700 and style normal set bold value, instead
of italics.

:QA Notes:
Test-cases: GT-18893

:Issues Addressed:
[PLAT-76009] The font differs from the expected result.

Change-Id: I4c7ad274393e81b55293b64175b19fc653b02a41
Originally-Reviewed-on: http://gpro.lgsvl.com/234643
Reviewed-on: http://gpro.lgsvl.com/234829
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>