[open][neva_upstreamable][media][ums] Update timestamp of WebMediaPlayerNeva on GetCurrentTime
:Release Notes:
Update timestamp of WebMediaPlayerNeva on GetCurrentTime

:Detailed Notes:
There are two types of timestamp that are related to video
progress. The first one is timestamp which is generated from
umediaserver. MediaPlayerUMS, UMediaClientImpl, UMediaServer share
this timestamp for video playback. The second one is timestamp
generated by interpolator in WebMediaPlayer(in this case,
WebMediaPlayerNeva), which moves the playback position cursor.

On this case, LoadCompleted() event is passed and media changes its
status to 'Play', and interpolator timer is started and playback
position cursor starts to progress. But the actual video is still on
buffering and the gap between playback position cursor and video
timestamp is made.

This patch eliminates returning code on MediaPlayerUMS and make the
two types of timers sync together.

:Testing Performed:
Tested on Raspberry Pi 3 webOS build
1. Run enact browser.
2. Go to:
   http://fileswp.lge.com/share/qa_managed_folder/mediaevents.html
3. Click the play button on the video, placed on the left side of the
   page.
4. Check that the progress cursor moves with video starts.

:QA Notes:
Test-cases: GT-27388

:Issues Addressed:
[PLAT-72761] [blink.html5] Constant Blank screen observed when playing the video

Change-Id: I35cf0e7babe34820b5a7ea8eb6c0f4779dc6c23e
Originally-Reviewed-on: http://gpro.lgsvl.com/237975
Reviewed-on: http://gpro.lgsvl.com/238791
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>