[open][neva_upstreamable][media][custom] Use WebOSMediaClient for camera
:Release Notes:
Use WebOSMediaClient (uMediaClient interface) for MediaPlayerCamera

:Detailed Notes:
Use common available uMediaClient interface (WebOSMediaClient)
for MediaPlayerCamera.
Prior to this MediaPlayerCamera was inheriting uMediaClient and
was using direct calls to UMS making many functions duplicate
and having separate implementation for uMediaClient.
With this patch we are using WebOSMediaClient for camera as well
So that there will be a single implementation of uMediaClient
for WebMediaPlayerNeva.

:Testing Performed:
Tested the camera test app and below test cases in raspberrypi3
build using the camera app changes as in patchset
http://gpro.lge.com/#/c/245779/ and the g-camera-pipeline patch
set http://gpro.lge.com/#/c/248138/

:QA Notes:
QR-1470, QR-1471, QR-1472, QR-1473, QR-1474

:Issues Addressed:
[PLAT-79764] Use  WebOSMediaClient for MediaPlayerCamera also

Change-Id: I06caa2c8d93d8758fe118c110db42cfbdf39c2ec
Reviewed-on: http://gpro.lge.com/245231
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Jose Dapena <jose.dapena@lge.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Tested-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>